pyproject.toml
[project]
name = "irispy-client-revised"
version = "0.0.1"
description = "A modified version of the Iris bot client in Python"
authors = [{ name = "ponyobot", email = "admin@ponyobot.kr" }]
requires-python = ">=3.9"
readme = "README.md"
dependencies = [
    "requests",
    "websockets",
    "pillow",
    "httpx"
]

[project.urls]
Original = "https://github.com/dolidolih/irispy-client"
Repository = "https://github.com/ponyobot/irispy-client-revised"

[project.scripts]
iris = "iris.cli:main"
